hermes: Create common Context object
Prior to this change, dependencies like the dbus::Bus and
objects needed to interact with the google-lpa library were passed to
object constructors as needed. As Hermes grows, with both the amount
of dependencies and number of classes increasing soon, it makes sense
to create a common context.

This change creates a Context singleton to serve this need. The choice
of a singleton was predicated on the desire to avoid the verbosity of
passing a Context to every object's constructor (assuming constructor
dependency injection), although either solution would work. Since
Daemon initializes the Context singleton, and the Daemon instance is
not destroyed until main() completes, lifetime mismatches between
Daemon and Context should not be a concern.

While Context::Initialize also would allow for a test framework to set
up the Context appropriately, integration testing in Hermes is impeded
by the fact that the Remote SIM Provisioning protocol is centered
around security and mutual authentication, making it difficult to set
up a fake flow and expect the test to behave analogously to
production.

BUG=chromium:1085825
TEST=All unit tests are passing.

Change-Id: Ice6c1402b6f8f8b26a837c86d64163f7d36f0ae9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2225176
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>