shill: Introduce method to migrate deprecated storage
Currently, a Service is only Save'd in special cases (mainly a Service
reaching a connected state, which will lead to Manager::PersistService
being triggered); in general a Service will have Load() called when its
corresponding Profile is pushed onto Manager's profile stack but will
not have Save() invoked. This makes it infeasible to deprecate
storage--in a reasonable amount of time--by migrating/deleting storage
elements in Save.

This change introduces a MigrateDeprecatedStorage method for Service
and friends (EapCredentials and VPNDriver) specifically for the
purpose of migrating or deleting storage elements.

BUG=chromium:1084279
TEST=All unit tests are passing.

Change-Id: I5f1491d3693135011a55f0eddbd38be2cbbd5563
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2229763
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>