nnapi: compile libraries with -fPIC option
Add //common-mk:pic to config for each static_library target to allow
them to be linked with the libneuralnetworks.so shared object library.

As noted in src/platform2/common-mk/BUILD.gn, -fPIC is the default
configuration for shared libraries, so is needed to link against
libneuralnetworks.so. See
https://stackoverflow.com/questions/53086636/linking-a-static-library-into-a-shared-library-without-fpic
for a detailed explanation of the bad things that happen without this
flag.

BUG=b:157391897
TEST=Linked libneuralnetworks.so from aosp-frameworks-ml-nn

Change-Id: I9501e2744010a42f0a30be9088242a8d46500b7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2224084
Commit-Queue: Alan Green <avg@chromium.org>
Tested-by: Alan Green <avg@chromium.org>
Reviewed-by: Stuart Langley <slangley@chromium.org>