dlcservice: Reorganize OnStatusUpdateAdvancedSignal
Reorganize OnStatusUpdateAdvancedSignal and HandleStatusResult to make
the flow linear and support error linking. Now that we are adding
metrics and we need to report the result of a DBus method call, having
diverging flows makes it very hard to report different results of a
function.

BUG=chromium:1069121
TEST=tast run $IP platform.DLCService

Change-Id: I9bf1dcdb61b0801c06241f1d0a5d66db5ba7395b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2226341
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>