dlcservice: Fix issue with untruncated writes
The only time we need a write to file method to not reduce/truncate the
size of the file is when we write into the dlc image themselves. All
other files including prefs, etc should truncate the file size if the
input data size is different than the existing file size.

BUG=None
TEST=unittests.

Change-Id: I862298e6ac6c0e4cd6461392415d59cd55b846cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2229338
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>