cryptohome: Rename AuthInput variables.
The AuthInput class instances are frequently referred to as
`user_input.` This is confusing because, in the Create() stage, the
AuthInput class contains data that is not input by the user at all. This
renames the instances to `auth_input`.

BUG=chromium:728125
TEST=FEATURES=test emerge-${BOARD} cryptohome; sign in and out.

Change-Id: Iec5c745b1934090ed608620969aecc2e3bb7908e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2226349
Tested-by: Greg Kerr <kerrnel@chromium.org>
Auto-Submit: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>