camera: Detach thread checker when the thread is stopped
Since we have tests (e.g. Camera3AlgoSandboxIPCErrorTest) which will
destroy Camera3Device first and then trying to connect again, it will
hit the DCHECK error since the thread checker is still attached to the
original thread.

Therefore, we should detach it from the thread if the thread is
stopped.

This CL also uses macros for thread checker to make it consistent to
other similar usages in this package.

BUG=b:158057598
TEST=Run cros_camera_test with DCHECK on

Change-Id: Ie03949d0e6a4104d6f0bad5c98a50e651b75bf71
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2228294
Tested-by: Wei Lee <wtlee@chromium.org>
Reviewed-by: Hung-yu Wu <hywu@chromium.org>
Commit-Queue: Wei Lee <wtlee@chromium.org>