shill: wifi: disallow Security!=SecurityClass at construction
Now that we've cleared up the storage and D-Bus concerns, to clarify
that no one ever constructs them with a non-SecurityClass value (i.e.,
it's "psk", not "wpa" or "rsn") and the unit tests are building more
realistic Endpoint and Service objects, let's finish the deed and
CHECK() if this is ever violated.

Clean up the last few unit tests that were doing the wrong thing.

BUG=b:157935328
TEST=unit tests; also suite:wifi_matfunc, etc.

Change-Id: Ib64a009a1af392fa49b9420331ea7a521c41c883
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2238462
Commit-Queue: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>