mtpd: Adjust seccomp filters to allow dup and dup3 syscalls
When a Pixel 4 (and from b/150071587 other Android models) is connected
to a Chromebook during boot the enumeration of its USB interface by
mtpd initially fails (in an underlying call to libmtp which uses
perror() to log the failure). In this code path, perror() uses dup()
which was previously blocked by mtpd's seccomp filters, killing mtpd
and leaving it zombied, meaning it is not restarted. This prevents
subsequent unplug / replug attempts of the device to fail.

Allowing dup() allows the failure to be cleanly logged, mtpd does not
die and when the device is put into File Transfer mode (or replugged)
it is detected as expected.

dup3 has also been added for future compatibility.

BUG=chromium:1090578
TEST=Emerged mtpd and deployed to nami, connected Pixel 4 phone,
     rebooted Chromebook, logged in, put Pixel 4 into File Transfer
     mode, Files.app appeared showing MTP contents as expected.

Change-Id: Ia2f69129f2f5fb8f4a807cf86ed3281a6b81578c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2230731
Tested-by: Josh Simmons <simmonsjosh@google.com>
Commit-Queue: Josh Simmons <simmonsjosh@google.com>
Reviewed-by: Sergei Datsenko <dats@chromium.org>