crash: Move filter_out logic entirely to main.
Most of the logic for deciding whether to handle a crash with
ArcCollector or UserCollector already lived in main, but some of it was
handled by each collector individually. Move it entirely to main and get
rid of the filter_out parameter.

This makes the code easier to follow and reduces the amount of tight
coupling between the arc collector and user collector, which will
simplify future refactoring.

BUG=chromium:1041929
TEST=cros_workon_make --test --board=eve  crash-reporter
TEST=tast run $DUT platform.UserCrash* arc.AppCrash*

Change-Id: I5a8907e93aef4d21f0ded61088be44bc05991c7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2233081
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>