diagnostics: debugd: Copy ProcessWithOutput into diagnostics
This class is useful to cros_healthd and in fact much of the code that
uses it in debugd may move there. We do not want it in brillo (see notes
on the bug).

Copy it into diagnostics, dropping support for not sandboxing.

BUG=chromium:1085464
TEST=emerge-nami diagnostics
Will also use pre-CQ

Change-Id: I9fdd5b82273911b438415c4e0a8367fd9c311aae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2236287
Commit-Queue: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>