dlcservice: Restructure periodic install check
Currently, the sole purpose of periodic install check is a back up in
case dlcservice does not receive update_engine's status through
signal. It is scheduled and cancelled constantly. This CL restructures
it a bit so we can always schedule such periodic check mechanism, but
only do the work (getting the status of update_engine through DBus
method) if we have not received a signal withing the lat 10 seconds.

Things done in this CL:

- Remove the need for getting the status of update_engine constantly for
each install attempt by caching the last received status instead.

- Introduces the global and cached update_engine status in the
system state. Once we have this in system state, we can let other parts
of the system like DlcManager or DlcBase to use it too.

- Let DlcBase::Purge() to decide whether to purge or not based on the
update_engine status. For example this function can decide to purge if
the DLC is not verified and hence it is not being updated.

- Use Clock object for refcount instead of base::Time::Now() so we can
easily test time based methods. Also improve the ref counting unittests
by using the Clock object.

- Same as above for DlcManager.

BUG=chromium:1085139
TEST=cros_workon_make --board reef --test dlcservice

Change-Id: I3b015dfeb88aebac5af4a60cf7b5dcfd7abc1d28
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2236725
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>