vm_tools: concierge: Account for disabled CPUs while starting a VM
This change adjusts the number of CPUs assigned to the VM to be upto the
CPUs desired by the client. As a VM is started, hyperthreading may be
disabled decreasing the number of available cores from the time the
client started the VM. Thus, the number of CPUs assigned to the VM
should be the min(available cores, cores requested by the client).

BUG=chromium:1076570
TEST=Run 'top' in crosh and 'penguin' shell. Ensure that CPUs match.

Change-Id: I98273a17fe20e928ce23c463f1dce0e2c1f93d3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2238529
Tested-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Auto-Submit: Abhishek Bhardwaj <abhishekbh@chromium.org>
Commit-Queue: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>