init: Don't remove specific VPD values on safe clobber-state
recovery_count is deprecated and not generated anywhere. There is a code
in factory to remove it, so there is no point keeping this around.

The original reason for deleting the first_active_omaha_ping_sent here
was that when we switched between normal and dev mode, we would have a
clean state. However, it doesn't seem that anyone ever cared about
this. This vpd value seems to be innocent and the more imporatnt reason
for not removing it ever is that it seems chromeos-startup sometimes
does a clobber-state on normal circumstances and that will remove this
value. If this value is removed from VPD, then we get wrong numbers for
first active pings and that is exactly what has been happening (it seems
like). This CL changes the logic so that this value is not removed ever
so it can be queried always by the update_engine and correct active ping
values can be sent always.

More informatino can be found in chromium:1071262 and chromium:912666.

BUG=chromium:1071262
TEST=CQ passes

Change-Id: I874082433588aa9405622c8af66571ef07799a60
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2238755
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Jong Ahn <jongahn@google.com>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>