vm_tools: 9s: change crate name to p9s for Rust 1.44.0
The version of Cargo shipped with Rust 1.44.0 no longer allows crate
names to start with a digit, so rename the crate to p9s.

The ebuild will be updated to install the binary as 9s, so nothing else
in the system needs to be updated.

BUG=chromium:1092719
TEST=emerge dev-rust/9s

Cq-Depend: chromium:2238871
Change-Id: I1e3479cd67ee5983f1c093b592acbe6972c7d36d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2238765
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>