crash: Log source directory in ephemeral collector
To debug a new flake in the linked bug, it's helpful to know which
directory the ephemeral collector is examining and copying from.

This statement should log once per boot, or if the ephemeral collector
is manually invoked (e.g. in the case of a mount/unmount failure on
startup or shutdown).

BUG=chromium:1039378
TEST=tast run $DUT crash.EarlyCrash*

Change-Id: Ie2d3d7885d3688b431c010a137723da3c44b6931
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2239073
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>