shill: key_file_store: Have DeleteKey return whether key was removed
Prior to this change, DeleteKey would return false if a key was
attempted to be removed from a non-existent group, but would return
true if a non-existent key was atempted to be removed from an existing
group. This change moves to the less awkward semantics of returning
true exclusively if the key was found and removed.

The only production code using this value currently is
PendingActivationStore, which itself has no production clients which
use its return value.

BUG=chromium:1084279
TEST=All unit tests are passing.

Change-Id: I78fbe2be06c5cc53a81008d15cc8d8524f8cf0f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2238960
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>