shill: key_file_store: Have DeleteGroup return whether group was removed
Prior to this change, DeleteGroup would always return true in
production code. This change causes DeleteGroup to return false if the
group did not exist (and hence was not removed) to be symmetric with
DeleteKey.

The only production code using this value currently is
Profile::AbandonService; this change removes the use of the return
value since the assumption was that the invocation would always return
true.

BUG=None.
TEST=All unit tests are passing.

Change-Id: I34ae1c13c0132e1269cf2b96336883dcd7b873db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2238961
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>