vm_tools: crostini_client: start lxd when starting VM
We are making "starting lxd" a separate step that need to be explicit
done through cicerone (and ultimately tremplin). This changes let
crostini_client start lxd when starting the VM.

A new struct `ProtobusSignalWatcher` is added to allow listening for
a signal before sending a request to avoid race condition.

BUG=chromium:1064512
TEST=vmc start termina (when vm is running or not)

Change-Id: I67a5c3f9ff792575e7f78f0cca1b92e1711c4449
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2237610
Tested-by: Jason Lin <lxj@google.com>
Commit-Queue: Jason Lin <lxj@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>