cryptohome: Remove kAsyncGuestMountTimer reporting.
This is dead code, because ReportTimerStop(kAsyncGuestMountTimer)
never gets called.

https://chromium.googlesource.com/chromiumos/platform2/+/d56ccf5e21ec8bffd06219c6bdd9e6a7f244e3e6
removed the AsyncMountGuest function. After that change, the only code
that ever mounts a Guest session is MountGuestEx.

But MountGuestEx will reply to the mount request with a BaseReply
object:
https://source.corp.google.com/chromeos_public/src/platform2/cryptohome/service.cc;l=2924

Which in turns sidesteps the if guard at:
https://source.corp.google.com/chromeos_public/src/platform2/cryptohome/service.cc;l=967

Which makes the histogram not report at:
https://source.corp.google.com/chromeos_public/src/platform2/cryptohome/service.cc;l=1009

Realistically, the entire if statement is dead code, but we can start
with this.

BUG=chromium:1091286
TEST=Compiles.

Change-Id: I56a6a14564f8eb3393d1bc605903fe1465c78b29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2239673
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Auto-Submit: Jorge Lucangeli Obes <jorgelo@chromium.org>