kerberos: add new value to seccomp policy files
Add |AF_INET6| value to kerberosd's seccomp policy files. This is a
tentative fix for the bug below. This failure hasn't been detected
before because the feature was not tested on a IPv6 network environment.
For more details, please refer to the discussion on the bug.

BUG=chromium:1092551
TEST=cros_workon_make --board=amd64-generic --test kerberos

Change-Id: Ic48bb969991e2ee22a82f8195d947eb53fcfafa1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2239946
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Tomasz Dobrowolski <tomdobro@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Felipe Andrade <fsandrade@chromium.org>
Commit-Queue: Felipe Andrade <fsandrade@chromium.org>