hermes: Introduce Euicc class
Currently the D-Bus API and internal code assume that there is a
single eUICC to communicate with: the one soldered onto the
board. This change provides the necessary framework for future
functionality to detect SIM slot changes and respond appropriately. In
particular, most of the previous Manager D-Bus API is now that of
Euicc, so that there is never any ambiguity about which eUICC a call
was meant for.

BUG=chromium:1085825, chromium:1091861
TEST=All unit tests are passing.

Change-Id: Idcd797b950e4c355541bc8c31379ba4db747c3dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2276588
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Pavan Holla <pholla@google.com>
Reviewed-by: Pavan Holla <pholla@google.com>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>