dlcservice: Add clock object SystemState
Introduce the use of Clock. This allows us to use real clock for
production code and a test clock in unittests so we can advance the
clock forward instead of sleeping to doing hacky job to reverse the
timestamps, etc. More CLs for using this will follow.

BUG=chromium:1085139
TEST=cros_workon_make --board reef --test dlcservice

Change-Id: I7a0facd33074f32611f14ad8b4bf7eb0ba880acd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2238824
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>