hermes: modem_qrtr: Get slot status on initialization
This change provides the functionality for ModemQrtr to request SIM
slot status and inform an EuiccManager of the received information. A
future change will provide indication functionality so that changes to
SIM slots can be updated without polling; currently this information
is only pulled on ModemQrtr initialization.

BUG=chromium:1085825
TEST=All unit tests are passing.

Change-Id: Ie8d148dd89ba587ad4bd657534e965170ef965ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2277199
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>