permission_broker: Log child path on error
In issue 1084507 we see that permission_broker is encountering an child
of a USB device which does not have a devtype. This change logs the
sysfs path of the child when this error is encountered so that we can
debug further.

BUG=chromium:1084507
TEST=Ran permission_broker unit tests

Change-Id: Ib1d38626438acdd73d359d057e805b98a5870827
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2240570
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Tested-by: Reilly Grant <reillyg@chromium.org>