login: Add switch for NNPalm json.
Adds a switch for passing NNPalm information through to chrome with a
JSON object as a switch value. The keys are passed through and can be
augmented in future.

Add a unit test to test behavior when:
1. No items are present in cros_config
2. both items are present in cros_config

which is what we expect.

BUG=chromium:1060051
TEST=Newly added unit test passes. Tested on Kohaku device with
intended chrome side code and things function perfectly.

Change-Id: I0c2e61c3cbfa8c61906b963e5ae85f8365be664e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2239812
Tested-by: Rob Schonberger <robsc@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Rob Schonberger <robsc@chromium.org>