diagnostics: Add device specific adapters for data.
Different devices expose information about themselves in various ways.
The adapters provide a unified interface to retrieve the required data
from the devices abstracting the caller from the implementation details.
Example of such a detail is a model of the device, which is represented
by different pseudo-files in sysfs.
So far the provided data is very limited, but will be extended in future cls.

BUG=b:151176984
TEST=unittest and `cros-health-tool telem --category=storage`

Change-Id: I1f84ce3fd11502744de73955ea2d9898db59f53a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2239814
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Paul Moy <pmoy@chromium.org>