sepolicy: Add schedutil rate_limit_us and make cpu_capacity RO
The upstream schedutil tunable is 'rate_limit_us' and so it should be
read-write. The cpu_capacity tunable was changed from read-write to
read-only in kernel version v5.2. So far we have no userspace code that
tries to write this, so make the selinux policy follow the default of
considering this a read-only attribute.

BUG=b:158331822
TEST=Cherrypick this CL into src/platform2/ followed by
`emerge-trogdor chromeos-base/selinux-policy && \
 cros deploy ${DUT_IP} chromeos-base/selinux-policy && \
 tast run ${DUT_IP} security.SELinuxFilesSystem`
 Watch the tast test pass.

Cq-Depend: chromium:2242573, chromium:2242572
Change-Id: I868f9b2ce3256f96a6548efd1c0c702f4da0f872
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2241571
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Joel Fernandes <joelaf@google.com>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>