cryptohome: Set finch userdataauth feature flag to be a kill switch
Previously userdataauth finch feature flag behaves like an override, if
it is on, it'll override the system's configuration. This CL change it to
a kill switch instead, that is, if it's off, it'll override, but if it's
on, it'll leave the decision to the config file on the system.

BUG=b:151062600
TEST=Manually reboot and check conf files

Change-Id: I74ca1ab2f384c97913dd8deed64ff3dd6db4ab35
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2241220
Tested-by: John L Chen <zuan@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Commit-Queue: John L Chen <zuan@chromium.org>