system-proxy: Add test HTTP server
This CL adds a simple HTTP server for testing that runs on a separate
thread and sends pre-set replies for HTTP requests.

The listening on the proxy url and port is blocking just for simplicity;
the HTTP server is not running on the main thread so we can't use a
base::FileDescriptorWatcher.

BUG=chromium:1042642
TEST=unittest

Change-Id: Ie502292cdc0e5421e60796b56050bda56634e347
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2245696
Tested-by: Andreea-Elena Costinas <acostinas@google.com>
Commit-Queue: Andreea-Elena Costinas <acostinas@google.com>
Reviewed-by: Pavol Marko <pmarko@chromium.org>