system-proxy: Send status code from the web server
System-proxy acts both as an HTTP client for the remote server and as
a server for the clients on Chrome OS.

Curl is designed as a client application. When failing to create a
tunnel, libcurl will ignore the body of the reply while the headers
will still contain "Content-length: x" which may cause errors
in clients on Chrome OS.

This CL changes the current implementation which forwards data blindly
in case of error to a response line that include the status code from
the remote web server and a custom reason-phrase that indicates
System-proxy failed to create the tunnel.

BUG=chromium:1042642
TEST=unittest

Change-Id: Ibb6dc5374fb29a65b842af5e77772e5c5118a5c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2245697
Tested-by: Andreea-Elena Costinas <acostinas@google.com>
Commit-Queue: Andreea-Elena Costinas <acostinas@google.com>
Reviewed-by: Pavol Marko <pmarko@chromium.org>