ippusb_bridge: add keep alive handler
Add an implementation of a keep alive handler to ippusb_bridge.
When the --keep-alive socket path is provided, ippusb_bridge will
monitor connections on that socket.  When a new client connects, a
thread will be spawned to service the client. If it reads a message
"\x0bkeep-alive\0" from the socket, it will notify the main loop to
postpone shutdown.

If no clients are connected and 10 seconds elapse without a keep alive
message, ippusb_bridge will cleanly shut down.

BUG=b:160720504
TEST=manually sent keep-alive pings with socat and check that they are
     received properly

Change-Id: Iff7c4fdd75b70d51981a3be283f80c506dc91f0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2280383
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Kalvin Lee <kdlee@chromium.org>