vm_tools: concierge: register Plugin VMs with permission service
To determine whether a Plugin VM can access microphone and/or camera,
concierge will register them with the new VM permission service
(accessible via DBus and residing in Chrome).

In the first phase we only support static permissions, with permissions
checked only upon VM startup; for the new settings to go into effect the
VM needs to be re-launched. Because if this restriction we do not need
to notify plugin of the access token or teach CRAS or Camera service of
the new permission service, but rather let concierge decide:

- whether to mount camera service socket into plugin jail
- whether to mount restricted CRAS socket or socket that allows capture
  into plugin jail

BUG=b:127478233,b:150341812
TEST=Run VM, toggled permissions on and off, observed them taking effect
after relaunching VM.

Cq-Depend: chrome-internal:3123455
Change-Id: I23ab81b96d5a298eb7da37749bfb8ef4f2396f1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2253190
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Jasmine Chen <lnishan@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>