cryptohome: Update update_userdataauth_from_features.sh to use Killswitch
IsCryptohomeUserDataAuthEnabled was added when the plan was to launch
UserDataAuth through finch, but now we intend to launch through normal
auto update channel instead, so we'll need a finch feature as killswitch
instead of launch switch.

To avoid using the same finch feature for different purpose, a new finch
feature is added and this script is changed to use that instead.

BUG=b:151062600
TEST=deploy with and without new chrome feature and check log.

Change-Id: Ibf861c929b31e37fdab1f21a601f2d00a316bf34
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2245156
Tested-by: John L Chen <zuan@chromium.org>
Commit-Queue: John L Chen <zuan@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>