sepolicy: Let tast tests r/w /run/crash_reporter
Crash-related tast tests need to access /run/crash_reporter to interact
with the system, but are not expressly permitted to do so.

This causes problems for selinux tests, since the test itself causes an
unintended selinux violation that masks the one we actually wanted.

NOTE- the DUT I'm running this on didn't ever display the problem in the
linked bug, so I can't be certain that it will fix it, but the test
should verify it's not totally broken.

BUG=chromium:1095177
TEST=deploy policy; run crash.SelinuxViolation.mock_consent

Change-Id: I11b223c9cc0862089c1de8c81c730d5cc734666a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2246310
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>