vm_tools: Test for borealis build using #if
crrev.com/c/2239072 adds the flag USE_VM_BOREALIS with the intention
that it will disable and enable borealis and termina specific
code. This flag is tested for using #ifdef and #ifndef, but in the
BUILD.gn file it is defined unconditionally, only its value
changes. This causes the build of maitred in termina to run as if it
were a part of borealis, with predictably bad consequences.

Fix this by using #if to test the value of USE_VM_BOREALIS instead of
its definedness.

BUG=chromium:1095280
BUG=chromium:1094864
TEST=Built and deployed a termina image

Change-Id: I522867e6e2d1ecc7ae86fc3ed7c22ff1c839d941
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2246130
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: Fergus Dall <sidereal@google.com>