smbfs: Enable setting of access and modification times
Extends SetAttr() so that the access and modification times of a file
or directory can be set. This allows modification time to be preserved
when copying files and directories from Chrome OS to an SMB share.

BUG=chromium:1093503
TEST=cros_run_unit_tests --board=nami --packages smbfs
TEST=Manually tested by copying a directory containing a file from the
DUT to a mounted SMB share; the (different) modification times of both
the directory and file it contained matched those on the DUT. Ensured
that truncate still worked by using truncate(1) on DUT command line.

Change-Id: I7ce17b4392b1719ba0afdb371eda619748c3a426
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2248004
Reviewed-by: Anand K Mistry <amistry@chromium.org>
Tested-by: Josh Simmons <simmonsjosh@google.com>
Commit-Queue: Josh Simmons <simmonsjosh@google.com>