vm_tools: Notify cicerone before VM shutdown
Currently concierge only notifies cicerone that a VM has stopped after
completely shutting it down. For stability metrics purposes, we need
to know before the shutdown signal is sent, so we can distinguish
between VM services stopping during a shutdown (which is expected) and
at other times (which is not).

BUG=chromium:977201
TEST=Shutdown VM and check that message is received

Change-Id: I3aa3eaa66aa201e823a6d2f595d209a47a277b86
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2263493
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>