system_api: add UNSPECIFIED to lorgnette enums
The enums in the lorgnette proto API did not have UNSPECIFIED values for
their 0 value.  Because there is no way to distinguish an unset field
from a 0 field in proto3, this means that our API is subtly broken.

Since nothing is depending on these enums yet, make a breaking proto API
change to add an UNSPECIFIED value as the 0 value in all enums.

BUG=b:156499700
TEST=emerge-${board} system_api

Change-Id: Ie89dfd03e2508ff61b5722087474c4946db36b83
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2250328
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>