vm_tools: Add service failure RPC to cicerone
This RPC allows the VM to report to the host when one of its services
has stopped. Cicerone will record this as a service failure in the
Crostini.Stability UMA histogram iff the VM should be in the normal
running state and the failed service is one cicerone recognises.

BUG=chromium:977201
TEST=Check that recorded events show up in chrome://histograms

Change-Id: I612d1ef178bc8af73ef706592da9c0c17e673b96
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2263494
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>