common-mk: update BASE_VER before it can be removed.
We want to remove this, or at least move it to aosp/external/libchrome.
Currently it's referred in many places as a fallback BASE_VER if
environment variable is not set (aka not inherit libchrome).

Let's keep it up to date, in sync with the uprev,  before it can be
removed.

BUG=chromium:1054279
TEST=CQ

Change-Id: I6b56610470de848cbac4e8577c3a7c2e66c071bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2251939
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Qijiang Fan <fqj@google.com>
Tested-by: Qijiang Fan <fqj@google.com>