arc-setup: Generate camera.prop every time
In theory, the prop file should be re-generated when the prop file
is older than arc-setup (or arc-apply-per-board-config in case of
ARCVM). It should be regenerated when the file is older than
/usr/share/arc[vm]/properties/build.prop too. Otherwise, changes
to e.g. build.prop won't be reflected to the generated file.

This CL does that by just removing the existing check of the prop
file. Since the prop file generation is almost instant, mtime check
of all these files is probably overkill.

BUG=None
TEST=corrupt camera.prop, sign out/in, verify camera.prop is
 generated again.

Change-Id: Ic0a26801d6fd532711481d769fbbce6d42a76be1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2253640
Tested-by: Yusuke Sato <yusukes@chromium.org>
Commit-Queue: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Ereth McKnight-MacNeil <ereth@chromium.org>