croslog: Add a message loop to main()
In the previous commit (crrev.com/c/2237609), a message loop was
not moved to main function in the command but in the test.

This CL adds a message loop in main() function in the command as
well as the test runner.

BUG=chromium:1072325
TEST=Tested manually on real device

Change-Id: I860ef9aa30d042beb006e6be56225a51f9d83fc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2256959
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Yoshiki Iguchi <yoshiki@chromium.org>
Tested-by: Yoshiki Iguchi <yoshiki@chromium.org>
Auto-Submit: Yoshiki Iguchi <yoshiki@chromium.org>