camera: libcamera_connector: Rework synchronization
This CL reworks the synchronization mechanisms in libcamera_connector.
The critical synchronization work is now centralized in CameraClient
which uses a lock to gate the opening/closing of cameras and the
returning of capture results.

BUG=b:151047930,b:159504861
TEST=Run cros_camera_connector_client and cros_camera_connector_test 100
iterations on eve and soraka with cros-debug.

Change-Id: Ic6791da2e031f7f14a696cfe918157de73718dca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2260178
Reviewed-by: Shik Chen <shik@chromium.org>
Tested-by: Jasmine Chen <lnishan@google.com>
Commit-Queue: Jasmine Chen <lnishan@google.com>
Auto-Submit: Jasmine Chen <lnishan@google.com>